Rename old tests
