+++ testold/e2e/namespaces_test.go
@@ -17,12 +17,12 @@ import (
 	"testing"
 
 	"github.com/nadundesilva/k8s-replicator/controllers"
-	"github.com/nadundesilva/k8s-replicator/test/utils/cleanup"
-	"github.com/nadundesilva/k8s-replicator/test/utils/controller"
-	"github.com/nadundesilva/k8s-replicator/test/utils/namespaces"
-	"github.com/nadundesilva/k8s-replicator/test/utils/resources"
-	"github.com/nadundesilva/k8s-replicator/test/utils/testdata"
-	"github.com/nadundesilva/k8s-replicator/test/utils/validation"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/cleanup"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/controller"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/namespaces"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/resources"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/testdata"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/validation"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/e2e-framework/pkg/envconf"
 	"sigs.k8s.io/e2e-framework/pkg/features"
