+++ testold/utils/controller/controller.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/nadundesilva/k8s-replicator/test/utils/cleanup"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/cleanup"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
