+++ testold/utils/namespaces/delete.go
@@ -16,7 +16,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/nadundesilva/k8s-replicator/test/utils/cleanup"
+	"github.com/nadundesilva/k8s-replicator/testold/utils/cleanup"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/e2e-framework/klient/k8s"
 	"sigs.k8s.io/e2e-framework/klient/wait"
